Added JSON support by brute-force copying a bunch of classes\n
The only existing class that is modified is AbstractCodeGenerator. A few more methods
were added so it could be used to represent both types of code generators from
an external caller.

The JSON support is currently not very sophisticated. It doesn't create object instances
for references (e.g. account.getOwner() returns the owner ID as string)

URLs are mapped to strings as opposed to java.net.URL with JPA. Jackson complained when
deserializing URL values that didn't conform to the URL spec (e.g. missing protocol)
so deserialization would break easily if imposing the java.net.URL type.

Currently, Jackson will throw away any fields in the JSON stream that do not have a mapping.
TBD whether this should continue to be default behavior, but it seems like a good idea to
keep parsing robust.

None of the copied code has tests yet (oops).